% Generated by roxygen2 (4.0.1): do not edit by hand
\name{expand}
\alias{expand}
\title{Expand column by splitting content with deliminter}
\usage{
expand(src, size, is_split, delim = ",")
}
\arguments{
\item{src}{List. Data to be splitted.}

\item{size}{integer. The size after splitted. Elements exceed the size will be discarded.}

\item{is_split}{logical. Indicate the corresponding column will be splitted or not.}

\item{delim}{char delimiter.}
}
\description{
Expand column by splitting content with deliminter
}
\details{
The size should not exceed the length after split or segment fault might occur.
}

