% Generated by roxygen2 (4.0.1): do not edit by hand
\name{splitSize}
\alias{splitSize}
\title{Evaluate the size after split}
\usage{
splitSize(src, delim)
}
\arguments{
\item{src}{character. Strings to be splitted.}

\item{delim}{char. Delimiter. Only the first char will be used.}
}
\value{
integer. The length of src splitted by delim.
}
\description{
Evaluate the size after split
}

